fix(bsb-otel): structure as entry-point DMZ to avoid premature MPI_Init
opentelemetry-instrument runs its target command in two Python phases:
entry-point discovery before `execl`, then sitecustomize after. When
`bsb_otel/__init__.py` imports `bsb.services` at module top, `MPI_Init`
fires in both phases — the first PMI registration consumes the SLURM
PMI2 slot, and the second fails with `PMI2_Init failed to initialize.
Return code: 14`. Manifests on CINECA g100 (slurm 23.11.8 / OpenMPI 4.1.1)
the moment `opentelemetry-instrument bsb compile ...` runs under srun.

Restructure `bsb_otel` so the package `__init__.py` and the modules
registered as entry points (`bsb_otel._otel_env`, `bsb_otel.exporters`)
contain no top-level `bsb*` imports — an explicit DMZ. The tracer,
context-vars, and lifecycle helpers move to a new `bsb_otel.tracer`
submodule that can freely import `bsb.services`; it only loads when
user code explicitly accesses it via a deep import, well after any
otel-instrument startup.

Callers across the monorepo switch from `from bsb_otel import X` to
deep imports like `from bsb_otel.tracer import get_bsb_tracer`. The
empty `__init__.py` enforces the DMZ rule structurally (no surface
on which to write a top-level `bsb` import).

Validated end-to-end on CINECA g100 from drodarie's \$SCRATCH workflow:
`opentelemetry-instrument bsb compile divergent_declive_column.yaml`
now clears MPI_Init on 48 ranks and proceeds into placement strategies
(previously failed at PMI2_Init for all ranks).
