+++ packages/bsb-otel/tests/test_bsb_otel.py
@@ -2,7 +2,7 @@ import unittest
 
 from opentelemetry import trace as ot
 
-from bsb_otel import BsbTracer, _tracer_registry, get_bsb_tracer
+from bsb_otel.tracer import BsbTracer, _tracer_registry, get_bsb_tracer
 from bsb_otel.testing import OTelFixture
